New style deps for stream-based modulemd 1.0.
This reverts the work done in b59869dbda515c1dd4a1ea40656013f2a26fb5d0
in order to regain the complexity that allows to specify the name and
the stream of the dependency.  We accomplished this before by
concatentating the name and the version into a single string, but the
possibility of dashes in the stream name means we can't reliably split
that string anymore.